requirements.txt
streamlit
groq
chromadb
sentence-transformers
pypdf
requests
python-dotenv
